Add a simulation code to OOS subperiod analysis
(1) Following Harry’s guideline, I created a simulation code as follows. For a single length-7 string, I draw a random number (0 or 1) seven times where the probability of seeing a 1 is ‘q’ for each dependent variable in Harry’s table. Each of the 7 draws is independent. Then I check if each string has at least one run of length k, where k = 1, 2, 3, 4, 5. For example, with one set of seven draws, we would have a string like “1101011”. Then we should record that this string has 1 run of length 1 and 2 runs of length 2. We draw 100,000 (or 1,000,000) sets of such strings and check what fraction of the total sets have at least one run of length k. 
(2)	The purpose of calculating this simulated fraction is to compare it with the probability of seeing at least one run of length k in a string of length 7, which is calculated by a function prob_of_run in Harry’s code. For example, prob_of_run(0.3,2,7) = 0.28 means that when a successful trial has a 30% chance, i.e. draw 1 30% of the time and 0 70% of the time, the probability of seeing at least one run of length 2 in a string of 7 trials is 28%.
(3)	I created a simulation function named frac_runs to calculate the fraction,  and then created another function compare_sim_data to make a table which compares the probabilities calculated by Harry’s function and the simulated fraction that I calculated.